ENH: Decompose fetchers.py into rocketpy/environment/fetchers package with dedicated submodules
