+++ rocketpy/environment/fetchers/base.py
@@ -1,0 +1,7 @@
+"""Base constants and logger for atmospheric data fetchers."""
+
+import logging
+
+logger = logging.getLogger(__name__)
+
+MAX_RETRY_DELAY_SECONDS = 600
